fsck.gfs2: Change bitmap_type variables to int
This patch changes all references of "q" from a uint8_t to a normal
int. This will allow us to check for bad return codes at some point.

+++ gfs2/fsck/lost_n_found.c
@@ -99,7 +99,7 @@ static void add_dotdot(struct gfs2_inode *ip)
 
 void make_sure_lf_exists(struct gfs2_inode *ip)
 {
-	uint8_t q;
+	int q;
 	struct dir_info *di;
 	struct gfs2_sbd *sdp = ip->i_sbd;
 	uint32_t mode;
